Add files via upload
This is  HACKER RANK question where we have print the sum of the given integer ,
but the twist is that it is not an integer and cannot covert to array.
********************************************************************************
Objective

The modulo operator, %, returns the remainder of a division. For example, 4 % 3 = 1 and 12 % 10 = 2. The ordinary division operator, /, returns a truncated integer value when performed on integers. For example, 5 / 3 = 1. To get the last digit of a number in base 10, use  as the modulo divisor.

Task

Given a five digit integer, print the sum of its digits.
Input Format

The input contains a single five digit number, n.

Constraints

100000<=n<=99999
Output Format

Print the sum of the digits of the five digit number.

Sample Input 

10564
Sample Output 

16

**********************************************************************************
EXPLANATION **

◉ The given integer is "" 10654""

◉ Start with a ""while loop "" where  ◆(int =! 0)

◉ We get the modulus (% 10 ) of int  and store it in a new int 

◉ Declare a new int SUM=0 and store the final value in it